Change item id from integer to long

+++ src/main/java/jenkins/advancedqueue/sorter/QueueItemCache.java
@@ -56,7 +56,7 @@ public class QueueItemCache {
 	}
 
 	// Keeps track of all items currently in the queue
-	private Map<Integer, ItemInfo> item2info = new HashMap<Integer, ItemInfo>();
+	private Map<Long, ItemInfo> item2info = new HashMap<Long, ItemInfo>();
 	// Keeps track of the last started item of the Job
 	private Map<String, ItemInfo> jobName2info = new HashMap<String, ItemInfo>();
 
@@ -70,7 +70,7 @@ public class QueueItemCache {
 	 * @return the {@link ItemInfo} for the provided id or <code>null</code> if the id is not in the
 	 *         queue
 	 */
-	synchronized public ItemInfo getItem(int itemId) {
+	synchronized public ItemInfo getItem(long itemId) {
 		return item2info.get(itemId);
 	}
 
@@ -85,13 +85,13 @@ public class QueueItemCache {
 	}
 
 	synchronized public ItemInfo addItem(ItemInfo itemInfo) {
-		Integer itemId = new Integer(itemInfo.getItemId());
+		Long itemId = new Long(itemInfo.getItemId());
 		item2info.put(itemId, itemInfo);
 		jobName2info.put(itemInfo.getJobName(), itemInfo);
 		return itemInfo;
 	}
 
-	synchronized public ItemInfo removeItem(int itemId) {
+	synchronized public ItemInfo removeItem(long itemId) {
 		return item2info.remove(itemId);
 	}
